+++ test/test_kps4.jl
@@ -1,8 +1,10 @@
 using Test, BenchmarkTools, StaticArrays, LinearAlgebra, KiteUtils
 using KiteModels, KitePodModels
 
+set_data_path(joinpath(dirname(dirname(pathof(KiteModels))), "data"))
+set = deepcopy(load_settings("system.yaml"))
 if ! @isdefined kcu
-    const kcu = KCU(se())
+    const kcu = KCU(set)
 end
 if ! @isdefined kps4
     const kps4 = KPS4(kcu)
@@ -47,7 +49,7 @@ function init_150()
     kps4.set.v_wind = 9.1
     kps4.set.mass = 6.21
     kps4.set.c_s = 0.6
-    kps4.set.damping = 473.0     # unit damping coefficient
+    kps4.set.damping = 473.0     # unit damping
     kps4.set.c_spring = 614600.0 # unit spring coefficent
     kps4.set.width = 4.9622
 end
@@ -151,7 +153,7 @@ end
     vel1 = KVec3(3.0, 4.0, 5.0)
     vel2 = KVec3(4.0, 5.0, 6.0)
     rho = kps4.set.rho_0
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         kps4.forces[i] .= zeros(3)
     end
     bytes = 0
@@ -159,7 +161,7 @@ end
         spring = kps4.springs[i]
         kps4.stiffness_factor = 0.5
         kps4.v_wind_tether .= KVec3(8.0, 0.1, 0.0)
-        bytes = @allocated KiteModels.calc_particle_forces!(kps4, pos1, pos2, vel1, vel2, spring, se().segments, se().d_tether/1000.0, rho, i)
+        bytes = @allocated KiteModels.calc_particle_forces!(kps4, pos1, pos2, vel1, vel2, spring, set.segments, set.d_tether/1000.0, rho, i)
     end
     # @test bytes == 0
     # Python output
@@ -174,7 +176,7 @@ end
          [ 12986.35257788861054    12986.7734548936750798  12986.8254733999201562]
          [ 23289.9810739697131794  23290.5422433097955945  23290.6116013181235758]
          [ -1883.1033393325606085  -1882.5421699924754648  -1882.4728119841502121]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test all(res[i,:] .≈ kps4.forces[i])
     end
 end
@@ -195,7 +197,7 @@ end
             [  77.4500000002498865    2.4810999999999996  134.1473350483994693]
             [  77.4500000002498865   -2.4810999999999996  134.1473350483994693]
            ]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         if ! all(pos1[i,:] .≈ pos[i])
             println(i, " ", pos1[i,:], " ", pos[i])
         else
@@ -205,7 +207,7 @@ end
 end
 
 function split_res(res)
-    particles = se().segments+KiteModels.KITE_PARTICLES
+    particles = set.segments+KiteModels.KITE_PARTICLES
     pos=res[1:3*particles]
     vel=res[3*particles+1:6*particles]
     pos, vel
@@ -213,8 +215,8 @@ end
 
 @testset "initial_residual      " begin
     init3()
-    res = zeros(MVector{6*(se().segments+KiteModels.KITE_PARTICLES), SimFloat})
-    X00 = zeros(SimFloat, 2*(se().segments+KiteModels.KITE_PARTICLES-1)+2)
+    res = zeros(MVector{6*(set.segments+KiteModels.KITE_PARTICLES), SimFloat})
+    X00 = zeros(SimFloat, 2*(set.segments+KiteModels.KITE_PARTICLES-1)+2)
     y0, yd0 = KiteModels.init(kps4, X00)
     residual!(res, yd0, y0, kps4, 0.0)
     res_pos, res_vel = split_res(res)
@@ -229,7 +231,7 @@ end
     init_150()
     kps4.set.elevation = 60.0
     kps4.set.profile_law = Int(EXP)
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         kps4.forces[i] .= zeros(3)
     end
     pos, vel, acc = KiteModels.init_pos_vel_acc(kps4, old=true, delta = 1e-6)
@@ -249,7 +251,7 @@ end
                [ 39.0565328469252862  -0.0092205388684263  47.4026845127173857]
                [  9.7854297156670409  84.4721933009509485   0.5027899265154456]
                [  9.7877962918810688 -84.5072746928884726   0.5210774922270867]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test isapprox(forces[i,:], kps4.forces[i], rtol=1e-4)
     end
 end
@@ -259,7 +261,7 @@ end
     init_150()
     kps4.set.elevation = 60.0
     kps4.set.profile_law = Int(EXP)
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         kps4.forces[i] .= zeros(3)
     end
     pos, vel, acc = KiteModels.init_pos_vel_acc(kps4, old=true, delta = 1e-6)
@@ -268,7 +270,7 @@ end
     alpha_depower = 0.1
     rel_steering = +0.1
     kps4.set.alpha_zero = 5.0
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         kps4.forces[i] .= zeros(3)
     end
     KiteModels.calc_aero_forces!(kps4, pos, vel, rho, alpha_depower, rel_steering)
@@ -283,7 +285,7 @@ end
             [ -81.3257203383301146   -2.0331316776625457 -454.15328374043645  ]
             [  -9.9385080719600989  -68.915067335201357    -0.9904916722429121]
             [ -11.4093091631036021   53.2874848115847612    0.7727700100708267]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test all(isapprox.(forces[i,:], kps4.forces[i], rtol=1e-4))
     end
 end
@@ -302,8 +304,8 @@ function init2()
     kps4.stiffness_factor = 0.5
     kps4.set.alpha = 1.0/7.0
     length = 150.0
-    kps4.segment_length = length/se().segments
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    kps4.segment_length = length/set.segments
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         kps4.forces[i] .= zeros(3)
     end
     KiteModels.init_springs!(kps4)
@@ -325,7 +327,7 @@ end
           [ 0.        0.        0.      ]
           [ 0.        0.        0.      ]
           [ 0.        0.        0.      ]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         @test all(kps4.res1[i] .≈ res1[i,:])
     end
     forces=[[ -1.1039795506226362  -0.0210281466470539   0.6374018106309699]
@@ -339,7 +341,7 @@ end
             [ 39.0565328469252862  -0.0092205388684263  47.4026845127173857]
             [  9.7854297156670409  84.4721933009509485   0.5027899265154456]
             [  9.7877962918810688 -84.5072746928884726   0.5210774922270867]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1 
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1 
         @test isapprox(forces[i,:], kps4.forces[i], rtol=1e-4)
     end
     res2 = [[  0.000001             0.000001             0.000001          ]
@@ -353,7 +355,7 @@ end
             [ 29.6664941261243911  -0.0070037210740637  45.816049670887935 ]
             [  9.9103998578748431  85.5509913012598417  10.3192110782116959]
             [  9.9127966577351092 -85.5865207191570505  10.3377321952086678]]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test isapprox(res2[i,:], kps4.res2[i], rtol=1e-4)
     end
 end
@@ -410,7 +412,7 @@ end
         150.                ,    0.                ][1:end-2]
     poss, vels=unpack(MVector{66, Float64}(y0s))
     @test length(pos) == length(poss)
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test all(pos[i] .≈ poss[i])
         @test all(vel[i] .≈ vels[i])
     end
@@ -437,7 +439,7 @@ end
 
     @test res2_[1, :] == res2[1]
     @test all(res2_[2, :] .≈ res2[2])
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         if ! all(res2_[i, :] .≈ res2[i])
             # println(i, res2_[i, :], " ", res2[i])
             @test_broken all(res2_[i, :] .≈ res2[i])
@@ -469,19 +471,19 @@ end
     kps4.v_wind .= kps4.v_wind_gnd * calc_wind_factor(kps4.am, height)
     res1, res2 = find_steady_state!(kps4; stiffness_factor=0.035, prn=false) 
     # TODO check why -9.81 appears in the residual
-    @test sum(res2) ≈ -9.81*(se().segments+ KiteModels.KITE_PARTICLES) # velocity and acceleration must be near zero
+    @test sum(res2) ≈ -9.81*(set.segments+ KiteModels.KITE_PARTICLES) # velocity and acceleration must be near zero
     pre_tension = KiteModels.calc_pre_tension(kps4)
     @test pre_tension > 1.0001
     @test pre_tension < 1.01
-    @test unstretched_length(kps4) ≈ 392.0              # initial, unstreched tether length
+    @test unstretched_length(kps4) ≈ 392.0              # initial, unstretched tether lenght
     # println("length: ", tether_length(kps4))
-    @test isapprox(tether_length(kps4), 408.74, rtol=1e-2) # real, streched tether length was: 406.4
+    @test isapprox(tether_length(kps4), 407.27, rtol=1e-2) # real, stretched tether length was: 406.4
 #    @test winch_force(kps) ≈ 276.25776695110034        # initial force at the winch [N]
 #    lift, drag = lift_drag(kps)
 #    @test lift ≈ 443.63303000106197                    # initial lift force of the kite [N]
 #    @test drag ≈ 94.25223134952152                     # initial drag force of the kite [N]
-#    @test lift_over_drag(kps) ≈ 4.706870316479931      # initlal lift-over-drag
-#    @test norm(v_wind_kite(kps)) ≈ 9.107670173739065   # inital wind speed at the height of the kite [m/s]
+#    @test lift_over_drag(kps) ≈ 4.706870316479931      # initial lift-over-drag
+#    @test norm(v_wind_kite(kps)) ≈ 9.107670173739065   # initial wind speed at the height of the kite [m/s]
 end
 
 @testset "test_spring_forces    " begin
@@ -493,7 +495,7 @@ end
     y0, yd0 = KiteModels.init(kps4)
     forces = spring_forces(kps4; prn=false)
     ref_forces = [3.928735076156923e-12, 3.928735076156923e-12, 3.928735076156923e-12, 0.0, -3.928735076156923e-12, 1.1786205228470769e-11, 2.160277004750972, 2.1602770047433926, 2.1602770047441373, 2.1602770047074573, 2.1602770047483513, 2.1602770047483513, 2.1602770047074573, 2.1602770047433926, 2.160277004685822]
-    for i in 1:se().segments + KiteModels.KITE_PARTICLES + 1
+    for i in 1:set.segments + KiteModels.KITE_PARTICLES + 1
         @test isapprox(forces[i], ref_forces[i], atol=1e-6, rtol=1e-4)
     end    
 end
@@ -539,11 +541,11 @@ end
 end
 
 @testset "Raptures" begin
-    kps4_ = KPS4(KCU(se()))
+    kps4_ = KPS4(KCU(set))
     integrator = KiteModels.init_sim!(kps4_; stiffness_factor=0.035, prn=false)
     kps4_.set.version = 2
     kps4_.stiffness_factor = 3
-    @test maximum(spring_forces(kps4_; prn=false)) > 20000
+    @test maximum(spring_forces(kps4_; prn=false)) > 9000
 end
 
 # TODO Add test for winch_force
@@ -561,6 +563,21 @@ end
     @test isdir("bin")
     cd("bin")
     @test isfile("create_sys_image")
+end
+
+@testset "turn" begin
+    # test the turn function
+    UPWIND_DIR = -pi/2 +deg2rad(10)
+
+    # y, yd
+    res=[1.0, 0, 1, 0, 0, 0]
+    res2 = KiteModels.turn(res, UPWIND_DIR)
+    @test res2[1] ≈ 0.984807753012208
+    @test res2[2] ≈ -0.17364817766693036
+    @test res2[3] ≈ 1.0
+    @test res2[4] ≈ 0.0
+    @test res2[5] ≈ 0.0
+    @test res2[6] ≈ 0.0
 end
 
 end