+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+### KiteModels v0.6.13 2024-12-06
+#### Changed
+- update the fields `set_steering`, `bearing` and `attractor` of the `SysState` struct 
+  in the function `update_sys_state!`
+- add the parameters `bearing` and `attractor` to the function `next_step!` for logging
+#### Fixed
+- fix #88: the function `init_sim!()` has the new parameter `upwind_dir` to define the
+  initial wind direction 
+
+### KiteModels v0.6.12 2024-12-01
+#### Changed
+- update the fields `set_torque`, `set_force`, `set_speed`, `alpha3`, `alpha4`, `roll`, `pitch`, `yaw`
+  of the `SysState` struct in the function `update_sys_state!`
+- add the parameter `set_force` to the function `next_step!` for logging
+- the four point kite model KPS4 was extended to include aerodynamic damping of pitch oscillations;
+  for this purpose, the parameters `cmq` and `cord_length` must be defined in `settings.yaml`
+- the four point kite model KPS4 was extended to include the impact of the deformation of the
+  kite on the turn rate; for this, the parameter `smc` must be defined in `settings.yaml`
+- improve examples
+- add the packages `JLD2` and `Colors` to the system image
+#### Added
+- add examples `calc_spectrum.jl` and `plot_spectrum` to plot the eigenfrequencies of the system
+- the script `calculate_rotational_inertia.jl` for calculating the inertia matrix of the kite
+- function `menu2()` which displays a menu with scripts for model verification
+
 ### KiteModels v0.6.11 2024-11-09
 #### Fixed
 - fixed bug in spring_forces(), it used 4000N hardcoded max force